All: minor
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4278 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ Generator/Generator/bboxIntersection.cpp
@@ -108,7 +108,7 @@ PyObject* K_GENERATOR::bboxIntersection(PyObject* self, PyObject* args)
     isIntersect = 1;
 
 #ifdef E_DOUBLEINT
-    return Py_BuildValue("l", long(isIntersect));
+    return Py_BuildValue("l", isIntersect);
 #else
     return Py_BuildValue("i", isIntersect);
 #endif
